Port EMFormatHTML to WebKit
EMFormatHTML is now able to display any kind of emails in WebKit webview.
Some functionaliy is solved by using JavaScript (height of iframe with
text/html part, automatic jump to beginning of email).
Embedded GtkWidgets (attachments bar, etc)  should work, but this feature
was not tested due to WebKit bug #63451.
This port is not completely asynchronous, additional resources (text/html
parts, embedded images) are being parsed and formatted synchronously (in
the main loop), because as of now WebKit does not provide any means to
change content of a resource asynchronously.