Port collapsable To/Cc/Bcc fields and collapsable headers to JavaScript
Use JavaScript to handle collapsing/expanding of To, Cc and Bcc header
fields and collapsing of the whole headers themselves.
The headers collapsing must notify "headers-state" property of EMFormatHTML
when changed. This is done by extending EWebView API by function for
installing callback functions to JavaScript.
Finally, the simple_headers property was removed from EMFormatHTML as it
was not used anymore.