Split functionality of EMFormat to "parse" and "write"
EMFormat is now able to traverse throught the CamelMimeMessage and
assign a EMFormatWriteFunc to each mime part. The mime part is stored
as EMFormatPURI struct, containing various metadata including the
write function, part ID and ptr to the part itself. Parts are stored
in mail_part_list (GList which determines the order in which parts will
be displayed) and mail_part_table (GHashTable which allows fast lookup
of parts by their ID).

When a content of particular part is requested, user can just find
the part by it's ID in mail_part_table and call the write_func which
will convert the part to HTML code ready to be displayed.

Alternatively, when a part represents an iCal event or anything else
that would be displayed as a GtkWidget, the widget_func is assigned
instead of write_func.

I tried to get rid of the way of passing informations between
parser functions through publicly accessible properties of EMFormat,
rather introducing EMFormatParserInfo structure passing as a parameter
to all EMFormatParserFunc. I think it's much cleaner.

The API of EMFormat is not final and most will be changed and polished
in future.

EMFormatQuote is only updated to build against the EMFormat, but most
probably does not work as it should.