Adapt EMFormat subclasses to the new EMFormat API
The EMFormat subclasses now work the same way EMFormat does -
providing EMFormatParseFunc for parsing the mail and EMFormatWriteFunc
for converting parts to displayable HTML code.

EMFormatHTML is almost entirely ported. Parsing and formatting
email as source code and analyzing and displaying attachments is
not working yet. All the headers-related code will be most probably
moved to EMFormatHMTLHeaders, as it is more then 600 lines of completely
independent code.

EMFormatHTMLDisplay is ported to the new API and partially works, but
attachments are not displayed and the functions need renaming to
conform better with the new functionality of EMFormat.

EMFormatHook is now adapted to the new API as well, but still needs
some more work.

The EMFormatHTMLPrint is only ported to compile against the new API,
but does not work yet.