Convert EMailDisplay to GtkScrolledWindow and change the way EMFormat is used.
Because embedding widgets directly into WebKit is broken, I decided to
split the email into multiple webview, thus allowing me to embbed
GtkWidgets between the webviews directly to a parent container.

The idea is to have EMailDisplay to implement most of the API of EWebView, thus
making it to look like a single EWebView from the outside. The EMailDisplay
would then work as a "proxy" resending the commands/signals/etc to all the
EWebViews and widgets it would manage.

This also requires completely another way of using EMFormat. Till now, it was
EMFormat-per-EMailDisplay. Since now, it's EMFormat-per-mail. This way, displaying
the same email in multiple views (the preview pane and EMailBrowser for example)
requires the email to be parsed only once and then the write_func of the required
mail parts are called. This can have positive impact on performance and user experience.

FIXME: For now, all created EMFormat objects remain stored in SoupSession and are not
removed even after the mail is not being displayed in any view. This is a big memleak
that needs to be fixed definitely.